Combine exportKey operations of ECDSA and ECDH
